apples to apples code samples for free text autograde (#257)
Now there are two sets of samples one set for partial points and one set
for full points. They are the same code all around and there are bash
and python versions for everything - with the new tabs. Also removed the
manual grading portion that was in free text auto-grade. Pretty sure
that didn't need to be there.